feat: Install vim-gtk3, instead of normal vim, for system clipboard builtin
In Neovim, "+y yanks your current selection to the system clipboard. In
regular old Vim, this behavior is *not* enabled by default, probably to
reduce attack surface area or binary size or something - but if you
install vim-gtk3, it is, as you can check with `vim --version | rg
clipboard`.

N.B., this does NOT get you access on Wayland. For that you probably
want something like

https://github.com/jasonccox/vim-wayland-clipboard/

which I am not yet including in Shell Bling Ubuntu.